pyproject.toml
[tool.poetry]
name = "scrapellm"
version = "0.1.0"
description = ""
authors = ["Mohit Kumar <mohitjandwani@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.9.0"
lxml = "^4.9.2"
beautifulsoup4 = "^4.12.2"
tiktoken = "^0.4.0"
openai = "^0.27.6"
pyppeteer = "^1.0.2"
asyncer = "^0.0.2"
streamlit = "^1.22.0"
markdownify = "^0.11.6"
scrapy = "^2.9.0"
unstructured = "^0.6.8"
requests = "2.28.0"
weaviate-client = "3.18.0"
pydantic = "^1.10.8"
tabulate = "^0.9.0"
tqdm = "^4.65.0"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^23.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
